Move InvestmentTransfers to Expense Namespace
