+++ RetireSimple.Engine/Data/Base/Investment.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 using RetireSimple.Engine.Data.Analysis;
+using RetireSimple.Engine.Data.Expense;
 using RetireSimple.Engine.Data.Investment;
 using RetireSimple.Engine.Data.User;
