+++ RetireSimple.Engine/Data/EngineDbContext.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore.Design;
 
 using RetireSimple.Engine.Data.Analysis;
 using RetireSimple.Engine.Data.Base;
+using RetireSimple.Engine.Data.Expense;
 using RetireSimple.Engine.Data.User;
 
 namespace RetireSimple.Engine.Data {
