+++ RetireSimple.Engine/Data/Expense/InvestmentTransfer.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 using RetireSimple.Engine.Data.Base;
 
-namespace RetireSimple.Engine.Data {
+namespace RetireSimple.Engine.Data.Expense {
 	public class InvestmentTransfer {
 		public int InvestmentTransferId { get; set; }
 		public int SourceInvestmentId { get; set; }
